+++ frontend/src/Components/EditTechnicalInput/PROSPTab.tsx
@@ -1,15 +1,14 @@
 import { Typography } from "@material-ui/core"
-import React, {
+import {
     Dispatch, MouseEventHandler, SetStateAction, useEffect, useState,
 } from "react"
 import styled from "styled-components"
 import { Button, Input, Label } from "@equinor/eds-core-react"
 import { Project } from "../../models/Project"
-import PROSPCaseList from "./PROSPCaseList"
 import { GetProspService } from "../../Services/ProspService"
 import { GetProjectService } from "../../Services/ProjectService"
 import { DriveItem } from "../../models/sharepoint/DriveItem"
-import PROSPCaseListNew from "./PROSPCaseListNew"
+import PROSPCaseList from "./PROSPCaseList"
 
 const ProspFieldWrapper = styled.div`
     margin-bottom: 2.5rem;
@@ -68,6 +67,7 @@ function PROSPTab({
             console.error("[PROSPTab] error while submitting form data", error)
         }
     }
+
     return (
         <div color="yellow">
             <TopWrapper color="danger">
@@ -83,15 +83,9 @@ function PROSPTab({
                 />
                 <Button variant="outlined" onClick={saveUrl}>Refresh</Button>
             </ProspFieldWrapper>
-
             <PROSPCaseList
                 project={project}
                 setProject={setProject}
-                driveItems={driveItems}
-            />
-
-            <PROSPCaseListNew
-                project={project}
                 driveItems={driveItems}
             />
